fix: extract same logic into unix::op
